chore: move UserVerifierInterface to core module
